# contracts/openapi/api.json
{
  "components": {
    "schemas": {
      "AbstentionReason": {
        "enum": [
          "insufficient_evidence",
          "conflicting_sources",
          "revision_mismatch"
        ],
        "title": "AbstentionReason",
        "type": "string"
      },
      "AnswerMode": {
        "enum": [
          "quick",
          "deep"
        ],
        "title": "AnswerMode",
        "type": "string"
      },
      "BddAnchor": {
        "additionalProperties": false,
        "properties": {
          "featureId": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Featureid",
            "type": "string"
          },
          "scenarioId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Scenarioid"
          },
          "stepId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Stepid"
          },
          "type": {
            "const": "bdd",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "featureId"
        ],
        "title": "BddAnchor",
        "type": "object"
      },
...
      "ChatTurnAccepted": {
        "additionalProperties": false,
        "description": "The durable identity returned after a turn has been accepted for processing.",
        "properties": {
          "chatId": {
            "title": "Chatid",
            "type": "string"
          },
          "state": {
            "const": "queued",
            "title": "State",
            "type": "string"
          },
          "turnId": {
            "title": "Turnid",
            "type": "string"
          }
        },
        "required": [
          "chatId",
          "turnId",
          "state"
        ],
        "title": "ChatTurnAccepted",
        "type": "object"
      },
      "ChatTurnRequest": {
        "additionalProperties": false,
        "description": "A browser request to create one turn in an existing chat.",
        "properties": {
          "answerMode": {
            "$ref": "#/components/schemas/AnswerMode",
            "default": "quick"
          },
          "clientRequestId": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Clientrequestid",
            "type": "string"
          },
          "message": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "requestedCorpusVersion": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestedcorpusversion"
          },
          "requestedEnvironment": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestedenvironment"
          },
          "resourceRefs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ResourceRef"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resourcerefs"
          },
          "sourceScope": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SourceFamily"
                },
                "maxItems": 4,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcescope"
          }
        },
        "required": [
          "clientRequestId",
          "message"
        ],
        "title": "ChatTurnRequest",
        "type": "object"
      },
      "CitationPreview": {
        "additionalProperties": false,
        "properties": {
          "anchor": {
            "$ref": "#/components/schemas/StructuralAnchor"
          },
          "chunkContentHash": {
            "maxLength": 71,
            "minLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Chunkcontenthash",
            "type": "string"
          },
          "citationId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Citationid",
            "type": "string"
          },
          "documentId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Documentid",
            "type": "string"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "prefix": {
            "default": "",
            "maxLength": 500,
            "title": "Prefix",
            "type": "string"
          },
          "quote": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Quote",
            "type": "string"
          },
          "revisionId": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Revisionid",
            "type": "string"
          },
          "sourceContentHash": {
            "maxLength": 71,
            "minLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Sourcecontenthash",
            "type": "string"
          },
          "suffix": {
            "default": "",
            "maxLength": 500,
            "title": "Suffix",
            "type": "string"
          }
        },
        "required": [
          "citationId",
          "documentId",
          "revisionId",
          "filename",
          "sourceContentHash",
          "chunkContentHash",
          "anchor",
          "quote"
        ],
        "title": "CitationPreview",
        "type": "object"
      },
      "CodeAnchor": {
        "additionalProperties": false,
        "properties": {
          "lineEnd": {
            "maximum": 2147483647,
            "minimum": 1,
            "title": "Lineend",
            "type": "integer"
          },
          "lineStart": {
            "maximum": 2147483647,
            "minimum": 1,
            "title": "Linestart",
            "type": "integer"
          },
          "path": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Path",
            "type": "string"
          },
          "repo": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Repo",
            "type": "string"
          },
          "symbol": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Symbol"
          },
          "type": {
            "const": "code",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "repo",
          "path",
          "lineStart",
          "lineEnd"
        ],
        "title": "CodeAnchor",
        "type": "object"
      },
      "ContentRole": {
        "enum": [
          "source",
          "generated_summary"
        ],
        "title": "ContentRole",
        "type": "string"
      },
...
      "ResourceMode": {
        "enum": [
          "required",
          "preferred",
          "scope"
        ],
        "title": "ResourceMode",
        "type": "string"
      },
      "ResourceRef": {
        "additionalProperties": false,
        "description": "Browser-provided retrieval intent; it cannot contain policy or ACL facts.",
        "properties": {
          "anchor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructuralAnchor"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "family": {
            "$ref": "#/components/schemas/SourceFamily"
          },
          "mode": {
            "$ref": "#/components/schemas/ResourceMode",
            "default": "preferred"
          },
          "requestedRevision": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Requestedrevision"
          },
          "sourceId": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Sourceid",
            "type": "string"
          }
        },
        "required": [
          "family",
          "sourceId"
        ],
        "title": "ResourceRef",
        "type": "object"
      },
      "RetrievalAnswerRequest": {
        "additionalProperties": false,
        "description": "Grounded-answer intent with the same narrowing-only search fields.",
        "properties": {
          "answerMode": {
            "$ref": "#/components/schemas/AnswerMode",
            "default": "quick"
          },
          "query": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "requestedCorpusVersion": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Requestedcorpusversion"
          },
          "requestedEnvironment": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Requestedenvironment"
          },
          "resourceRefs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ResourceRef"
                },
                "maxItems": 20,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Resourcerefs"
          },
          "sources": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SourceFamily"
                },
                "maxItems": 4,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Sources"
          },
          "topK": {
            "anyOf": [
              {
                "maximum": 100,
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Topk"
          }
        },
        "required": [
          "query"
        ],
        "title": "RetrievalAnswerRequest",
        "type": "object"
      },
      "RetrievalAnswerResponse": {
        "additionalProperties": false,
        "properties": {
          "abstained": {
            "title": "Abstained",
            "type": "boolean"
          },
          "abstentionReason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AbstentionReason"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/RetrievalCitation"
            },
            "title": "Citations",
            "type": "array"
          },
          "claims": {
            "items": {
              "$ref": "#/components/schemas/RetrievalClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "contextSnapshotId": {
            "minLength": 1,
            "title": "Contextsnapshotid",
            "type": "string"
          },
          "corpusVersion": {
            "minLength": 1,
            "title": "Corpusversion",
            "type": "string"
          },
          "degradationReasons": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Degradationreasons"
          },
          "degradedMode": {
            "title": "Degradedmode",
            "type": "boolean"
          },
          "queryPlanId": {
            "minLength": 1,
            "title": "Queryplanid",
            "type": "string"
          },
          "retrievalProfileId": {
            "minLength": 1,
            "title": "Retrievalprofileid",
            "type": "string"
          },
          "traceId": {
            "minLength": 1,
            "title": "Traceid",
            "type": "string"
          }
        },
        "required": [
          "traceId",
          "queryPlanId",
          "contextSnapshotId",
          "corpusVersion",
          "retrievalProfileId",
          "degradedMode",
          "answer",
          "abstained",
          "claims",
          "citations"
        ],
        "title": "RetrievalAnswerResponse",
        "type": "object"
      },
      "RetrievalCitation": {
        "additionalProperties": false,
        "properties": {
          "chunkContentHash": {
            "maxLength": 71,
            "minLength": 71,
            "pattern": "^sha256:[0-9a-f]{64}$",
            "title": "Chunkcontenthash",
            "type": "string"
          },
          "chunkId": {
            "minLength": 1,
            "title": "Chunkid",
            "type": "string"
          },
          "citationId": {
            "minLength": 1,
            "title": "Citationid",
            "type": "string"
          },
          "contentRole": {
            "$ref": "#/components/schemas/ContentRole"
          },
          "derivedFromChunkIds": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Derivedfromchunkids"
          },
          "evidenceLabel": {
            "minLength": 1,
            "title": "Evidencelabel",
            "type": "string"
          },
          "logicalChunkId": {
            "minLength": 1,
            "title": "Logicalchunkid",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/RetrievalSourceRevision"
          }
        },
        "required": [
          "citationId",
          "evidenceLabel",
          "chunkId",
          "logicalChunkId",
          "source",
          "chunkContentHash",
          "contentRole"
        ],
        "title": "RetrievalCitation",
        "type": "object"
      },
      "RetrievalClaim": {
        "additionalProperties": false,
        "properties": {
          "answerEnd": {
            "maximum": 2147483647,
            "minimum": 0,
            "title": "Answerend",
            "type": "integer"
          },
          "answerStart": {
            "maximum": 2147483647,
            "minimum": 0,
            "title": "Answerstart",
            "type": "integer"
          },
          "citationIds": {
            "items": {
              "type": "string"
            },
            "title": "Citationids",
            "type": "array"
          },
          "claimId": {
            "minLength": 1,
            "title": "Claimid",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "claimId",
          "text",
          "answerStart",
          "answerEnd",
          "citationIds"
        ],
        "title": "RetrievalClaim",
        "type": "object"
      },
...
      "SourceFamily": {
        "enum": [
          "doc",
          "code",
          "bdd",
          "failure"
        ],
        "title": "SourceFamily",
        "type": "string"
      },
      "StructuralAnchor": {
        "description": "A closed, structural location inside one authorized source family.",
        "discriminator": {
          "mapping": {
            "bdd": "#/components/schemas/BddAnchor",
            "code": "#/components/schemas/CodeAnchor",
            "document": "#/components/schemas/DocumentAnchor",
            "failure": "#/components/schemas/FailureAnchor",
            "openapi": "#/components/schemas/OpenApiAnchor"
          },
          "propertyName": "type"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/DocumentAnchor"
          },
          {
            "$ref": "#/components/schemas/CodeAnchor"
          },
          {
            "$ref": "#/components/schemas/BddAnchor"
          },
          {
            "$ref": "#/components/schemas/OpenApiAnchor"
          },
          {
            "$ref": "#/components/schemas/FailureAnchor"
          }
        ],
        "title": "StructuralAnchor"
      }
    }
  },
  "info": {
    "title": "TAP API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/chats/{chat_id}/turns": {
      "post": {
        "description": "Reserve the public route until the durable turn workflow is implemented.",
        "operationId": "chat_create_turn",
        "parameters": [
          {
            "in": "path",
            "name": "chat_id",
            "required": true,
            "schema": {
              "title": "Chat Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatTurnAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Request validation failed"
          },
          "501": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Turn workflow not implemented"
          }
        },
        "summary": "Create Chat Turn"
      }
    },
    "/v1/citations/{citation_id}": {
      "get": {
        "operationId": "citation_get_preview",
        "parameters": [
          {
            "in": "path",
            "name": "citation_id",
            "required": true,
            "schema": {
              "title": "Citation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitationPreview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid citation ID"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599,
                      "minimum": 100,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Knowledge runtime unavailable"
          }
        },
        "summary": "Get Citation",
        "tags": [
          "knowledge"
        ]
      }
    },
    "/v1/knowledge/answers": {
      "post": {
        "operationId": "knowledge_create_answer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalAnswerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalAnswerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 100.0,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Invalid answer request"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": false,
                  "description": "RFC 9457 problem details returned by the public HTTP interface.",
                  "properties": {
                    "detail": {
                      "minLength": 1,
                      "title": "Detail",
                      "type": "string"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Instance"
                    },
                    "status": {
                      "maximum": 599.0,
                      "minimum": 100.0,
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "minLength": 1,
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "pattern": "^https://",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "detail"
                  ],
                  "title": "ProblemDetails",
                  "type": "object"
                }
              }
            },
            "description": "Knowledge runtime unavailable"
          }
        },
        "summary": "Create Answer",
        "tags": [
          "knowledge"
        ]
      }
    },
...
  }
}